Quiet a couple of warnings: use an explicit cast on a safe type conversion (wider->narrower integer) and #include prototypes for a couple of exported functions.

+++ src/na/na_ucx.c
@@ -522,7 +522,7 @@ na_ucx_addr_hash(hg_hash_table_key_t key)
 
     /* Shift all bits out of `mix`, XOR into the hash code. */
     for (code = 0; mix != 0; mix /= shiftout)
-        code ^= (mix % shiftout);
+        code ^= (unsigned int)(mix % shiftout);
 
     return code;
 }
